Moved relevant methods to AWS_cloud class and adjusted the rest in config and main_controller
