+++ MicroPython/src/cloud/KAA_cloud.py
@@ -1,0 +1,9 @@
+from cloud.cloud_interface import CloudProvider
+
+
+class KAA_cloud(CloudProvider):
+    def __init__(self) -> None:
+        pass
+
+    def device_configuration(self, a, b):
+        return a
