+++ MicroPython/src/controller/main_controller.py
@@ -6,6 +6,7 @@ import machine
 import utime
 from cloud.AWS_cloud import AWS_cloud
 from cloud.cloud_interface import CloudProvider, Providers
+from cloud.KAA_cloud import KAA_cloud
 from common import config, utils
 from communication import wirerless_connection_controller
 from data_acquisition import data_acquisitor
@@ -46,8 +47,8 @@ class MainController:
         self.cloud_provider = self.get_cloud_provider()
 
         web_app.setup(get_measurement_hook=self.get_measurement,
-                      configure_device_hook=self.device_configuration,
-                      configure_aws_hook=self.configure_data_from_terraform,
+                      configure_device_hook=self.cloud_provider.device_configuration,
+                      configure_aws_hook=self.cloud_provider.configure_data_from_terraform,
                       configure_sensor_hook=self.configure_sensor,
                       start_test_data_acquisition=self.start_test_data_acquisition_hook,
                       start_data_acquisition=self.start_data_acquisition_hook,
@@ -91,7 +92,7 @@ class MainController:
             return AWS_cloud()
 
         elif config.cfg.cloud_provider == Providers.KAA:
-            pass
+            return KAA_cloud()
 
     def process_event(self, event: MainControllerEvent) -> None:
         """
@@ -101,7 +102,8 @@ class MainController:
         """
         if event.event_type == MainControllerEventType.CONFIGURE_ACCESS_POINT:
             logging.debug("Processing configure access point event")
-            access_point_name = "{}_{}".format(ACCESS_POINT_BASE_NAME, config.cfg.device_uid)
+            access_point_name = "{}_{}".format(
+                ACCESS_POINT_BASE_NAME, config.cfg.device_uid)
             logging.debug("Access point name {}".format(access_point_name))
 
             self.configure_access_point()
@@ -188,82 +190,85 @@ class MainController:
         """
         return self.data_acquisitor.get_single_measurement()
 
-    def device_configuration(self, data: dict) -> int:
-        """
-        Configures device in the cloud. Function used as hook to web_app.
-        :param data: parameters to connect to wifi.
-        :return: Error code (0 - OK, 1 - Error).
-        """
-        ssid = data['ssid']
-        password = data['password']
-        config.cfg.password = password
-        config.cfg.ssid = ssid
-        config.cfg.save()
-        logging.info("Wifi config. Wifi ssid {} Wifi password {}".format(ssid, password))
+    # TODO: DELETE AND MOVE TO AWS_cloud
+    # def device_configuration(self, data: dict) -> int:
+    #     """
+    #     Configures device in the cloud. Function used as hook to web_app.
+    #     :param data: parameters to connect to wifi.
+    #     :return: Error code (0 - OK, 1 - Error).
+    #     """
+    #     ssid = data['ssid']
+    #     password = data['password']
+    #     config.cfg.password = password
+    #     config.cfg.ssid = ssid
+    #     config.cfg.save()
+    #     logging.info("Wifi config. Wifi ssid {} Wifi password {}".format(ssid, password))
 
-        wireless_controller = wirerless_connection_controller.get_wireless_connection_controller_instance()
-        try:
-            utils.connect_to_wifi(wireless_controller)
-            logging.info(wireless_controller.sta_handler.ifconfig())
-            self.configure_aws_thing()
-        except Exception as e:
-            logging.error("Exception catched: {}".format(e))
-            event = MainControllerEvent(MainControllerEventType.ERROR_OCCURRED)
-            self.add_event(event)
-            return 1
+    #     wireless_controller = wirerless_connection_controller.get_wireless_connection_controller_instance()
+    #     try:
+    #         utils.connect_to_wifi(wireless_controller)
+    #         logging.info(wireless_controller.sta_handler.ifconfig())
+    #         self.configure_aws_thing()
+    #     except Exception as e:
+    #         logging.error("Exception catched: {}".format(e))
+    #         event = MainControllerEvent(MainControllerEventType.ERROR_OCCURRED)
+    #         self.add_event(event)
+    #         return 1
 
-        config.cfg.ap_config_done = True
-        config.cfg.save()
-        machine.reset()
-        return 0
+    #     config.cfg.ap_config_done = True
+    #     config.cfg.save()
+    #     machine.reset()
+    #     return 0
 
-    def configure_aws_thing(self) -> bool:
-        """
-        Register ESP as thing in AWS cloud.
-        :return: Error code (True - OK, False - Error).
-        """
-        logging.info("Allocated bytes on heap before gc {}".format(gc.mem_alloc()))
-        gc.collect()
-        logging.info("Allocated bytes on heap after gc {}".format(gc.mem_alloc()))
-        self.configure_data_from_terraform()
-        logging.debug("Authorization request...")
-        jwt_token = config.ESPConfig.authorization_request()
+    # TODO: DELETE AND MOVE TO AWS_cloud
+    # def configure_aws_thing(self) -> bool:
+    #     """
+    #     Register ESP as thing in AWS cloud.
+    #     :return: Error code (True - OK, False - Error).
+    #     """
+    #     logging.info("Allocated bytes on heap before gc {}".format(gc.mem_alloc()))
+    #     gc.collect()
+    #     logging.info("Allocated bytes on heap after gc {}".format(gc.mem_alloc()))
+    #     self.configure_data_from_terraform()
+    #     logging.debug("Authorization request...")
+    #     jwt_token = config.ESPConfig.authorization_request()
 
-        if jwt_token is not "":
-            cert_dict = config.ESPConfig.configuration_request(jwt_token)
-            config.ESPConfig.save_certificates(cert_dict)
-            logging.debug("Configuration done")
-            return True
-        else:
-            logging.error("Problem with authorization")
-            event = MainControllerEvent(MainControllerEventType.ERROR_OCCURRED)
-            self.add_event(event)
-            return False
+    #     if jwt_token is not "":
+    #         cert_dict = config.ESPConfig.configuration_request(jwt_token)
+    #         config.ESPConfig.save_certificates(cert_dict)
+    #         logging.debug("Configuration done")
+    #         return True
+    #     else:
+    #         logging.error("Problem with authorization")
+    #         event = MainControllerEvent(MainControllerEventType.ERROR_OCCURRED)
+    #         self.add_event(event)
+    #         return False
 
-    @staticmethod
-    def configure_data_from_terraform() -> None:
-        """
-        Setup data from terraform to connect to AWS.
-        :return: None
-        """
-        logging.debug("configure_data_from_terraform")
-        aws_configuration = config.cfg.load_aws_config_from_file()
-        if 'aws_iot_endpoint' in aws_configuration.keys():
-            config.cfg.aws_endpoint = aws_configuration['aws_iot_endpoint'].get("value").get("endpoint_address")
+    # TODO: DELETE AND MOVE TO AWS_cloud
+    # @staticmethod
+    # def configure_data_from_terraform() -> None:
+    #     """
+    #     Setup data from terraform to connect to AWS.
+    #     :return: None
+    #     """
+    #     logging.debug("configure_data_from_terraform")
+    #     aws_configuration = config.cfg.load_aws_config_from_file()
+    #     if 'aws_iot_endpoint' in aws_configuration.keys():
+    #         config.cfg.aws_endpoint = aws_configuration['aws_iot_endpoint'].get("value").get("endpoint_address")
 
-        if 'visualization_url' in aws_configuration.keys():
-            visualization_url = aws_configuration['visualization_url'].get("value")
-            config.cfg.api_url = 'https://' + visualization_url + '/api/'
+    #     if 'visualization_url' in aws_configuration.keys():
+    #         visualization_url = aws_configuration['visualization_url'].get("value")
+    #         config.cfg.api_url = 'https://' + visualization_url + '/api/'
 
-        if 'esp_login' in aws_configuration.keys():
-            config.cfg.api_login = aws_configuration['esp_login'].get("value")
+    #     if 'esp_login' in aws_configuration.keys():
+    #         config.cfg.api_login = aws_configuration['esp_login'].get("value")
 
-        if 'esp_password' in aws_configuration.keys():
-            config.cfg.api_password = aws_configuration['esp_password'].get("value")
+    #     if 'esp_password' in aws_configuration.keys():
+    #         config.cfg.api_password = aws_configuration['esp_password'].get("value")
 
-        config.cfg.save()
+    #     config.cfg.save()
 
-        logging.debug("Configure data from terraform ends")
+    #     logging.debug("Configure data from terraform ends")
 
     @staticmethod
     def configure_sensor(sensor_configuration: dict) -> None:
@@ -275,7 +280,8 @@ class MainController:
         logging.debug("Configure sensor")
         print(sensor_configuration)
         if 'publishing_period_ms' in sensor_configuration.keys():
-            config.cfg.data_publishing_period_in_ms = int(sensor_configuration['publishing_period_ms'])
+            config.cfg.data_publishing_period_in_ms = int(
+                sensor_configuration['publishing_period_ms'])
         if 'dht_type' in sensor_configuration.keys():
             config.cfg.dht_type = sensor_configuration['dht_type']
 
@@ -300,39 +306,43 @@ class MainController:
         """
         Configure AP to get wifi ssid and password
         """
-        access_point_name = "{}_{}".format(ACCESS_POINT_BASE_NAME, config.cfg.device_uid)
+        access_point_name = "{}_{}".format(
+            ACCESS_POINT_BASE_NAME, config.cfg.device_uid)
         wireless_controller = wirerless_connection_controller.get_wireless_connection_controller_instance()
-        wireless_controller.configure_access_point(access_point_name, "password")
+        wireless_controller.configure_access_point(
+            access_point_name, "password")
         self.web_server_thread = _thread.start_new_thread(web_app.run, [])
 
     @staticmethod
     def test_connection_with_wifi_and_cloud() -> None:
-        wireless_controller, mqtt_communicator = utils.get_wifi_and_aws_handlers(sync_time=True)
+        wireless_controller, mqtt_communicator = utils.get_wifi_and_aws_handlers(
+            sync_time=True)
 
         mqtt_communicator.disconnect()
         wireless_controller.disconnect_station()
 
-    def publish_data(self) -> None:
-        """
-        Publish data to the cloud
-        """
-        wireless_controller, mqtt_communicator = utils.get_wifi_and_aws_handlers(sync_time=False)
+    # TODO: DELETE AND MOVE TO AWS_cloud
+    # def publish_data(self) -> None:
+    #     """
+    #     Publish data to the cloud
+    #     """
+    #     wireless_controller, mqtt_communicator = utils.get_wifi_and_aws_handlers(sync_time=False)
 
-        certificates_existence, *_ = config.ESPConfig.read_certificates()
-        if not certificates_existence:
-            logging.debug("No AWS Certificates, configure_aws_thing()")
-            self.configure_aws_thing()
+    #     certificates_existence, *_ = config.ESPConfig.read_certificates()
+    #     if not certificates_existence:
+    #         logging.debug("No AWS Certificates, configure_aws_thing()")
+    #         self.configure_aws_thing()
 
-        logging.debug("data to send = {}".format(self.data_acquisitor.data))
-        logging.info(config.cfg.aws_topic)
-        result = mqtt_communicator.publish_message(payload=self.data_acquisitor.data, topic=config.cfg.aws_topic,
-                                                   qos=config.cfg.QOS)
+    #     logging.debug("data to send = {}".format(self.data_acquisitor.data))
+    #     logging.info(config.cfg.aws_topic)
+    #     result = mqtt_communicator.publish_message(payload=self.data_acquisitor.data, topic=config.cfg.aws_topic,
+    #                                                qos=config.cfg.QOS)
 
-        if not result:
-            logging.error("Error publishing data to MQTT in send_data()")
+    #     if not result:
+    #         logging.error("Error publishing data to MQTT in send_data()")
 
-        mqtt_communicator.disconnect()
-        wireless_controller.disconnect_station()
+    #     mqtt_communicator.disconnect()
+    #     wireless_controller.disconnect_station()
 
     @staticmethod
     def print_time() -> None:
